use the treeview jquery code in a react app
this will allow us to stop using the angularjs navbar
