+++ treeview/themes/default/images/style.css
@@ -15,7 +15,7 @@
 
 #parent {
     overflow:hidden;
-    margin-top: 1em;
+    padding-top: 10px;
 }
 
 #filterDropDown {
@@ -24,6 +24,19 @@
 
 #searchDiv {
     overflow: hidden;
+}
+
+#anatomyHeading {
+  display: 'none';
+  margin-left: 6px;
+}
+#filterDropDown {
+  visibility: 'hidden';
+}
+
+.side-panel-container {
+  height: 100%;
+  overflow-y: auto;
 }
 
 @media screen and (max-width: 830px){
@@ -42,8 +55,8 @@
     border-left:none;
 }
 
-.btn{
-    border-radius:0;
+.btn {
+    border-radius:0 !important;
 }
 
 .input-btn {
@@ -94,9 +107,6 @@ a.facet-heading-link:hover {
     display: none;
 }
 
-.jstree-grid-wrapper{
-  overflow: hidden!important;
-}
 .back-to-top {
     background: none;
     margin: 0;
@@ -2027,4 +2037,24 @@ input {
     background-position: -130px -4px !important;
     background-image: url(32px.png) !important;
   }
+}
+
+/* MISSING FROM BOOTSTRAP 5: */
+.panel-heading {
+  padding: 10px 15px;
+  border-bottom: 1px solid transparent;
+  border-top-left-radius: 3px;
+  border-top-right-radius: 3px;
+}
+
+.panel-default > .panel-heading {
+  color: #333;
+  background-color: #f5f5f5;
+  border-color: #ddd;
+}
+
+.btn.btn-primary.btn-inverted[disabled] {
+  color: #ccc !important;
+  border-color: #ccc !important;
+  background-color: white !important;
 }
