+++ tsconfig.json
@@ -10,7 +10,7 @@
       "DOM",
       "ESNext"
     ],
-      // "allowJs": true,                             /* Allow javascript files to be compiled. */
+    "allowJs": true,                             /* Allow javascript files to be compiled. */
     // "checkJs": true,                             /* Report errors in .js files. */
     "jsx": "react-jsx",                           /* Specify JSX code generation: 'preserve', 'react-native', 'react', 'react-jsx' or 'react-jsxdev'. */
     // "declaration": true,                            /* Generates corresponding '.d.ts' file. */
